refactor(boards): refactor out new board asset (itemRenderer)
