+++ packages/components/src/simulation-assets/index.ts
@@ -1,4 +1,5 @@
-export * from './create-items';
+export * from './items';
+export * from './expandable-item-renderer';
 export * from './item-renderer';
 export * from './li-item-renderer';
 export * from './stateful-item-renderer';
