+++ packages/components/src/simulation-assets/item-renderer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { ListItemProps } from '../list/list';
 import { SearchableText } from '../searchable-text/searchable-text';
-import type { ItemData } from './create-items';
+import type { ItemData } from './index';
 import { classes, st } from './item-renderer.st.css';
 
 export const ItemRenderer: React.FC<ListItemProps<ItemData>> = (props) => {
