+++ packages/components/src/simulation-assets/li-item-renderer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { ListItemProps } from '../list/list';
-import type { ItemData } from './create-items';
+import type { ItemData } from './index';
 import { classes, st } from './item-renderer.st.css';
 
 export const LIItemRenderer: React.FC<ListItemProps<ItemData>> = (props) => {
