+++ packages/components/src/simulation-assets/stateful-item-renderer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { ListItemProps } from '../list/list';
-import type { ItemData } from './create-items';
+import type { ItemData } from './';
 import { classes, st } from './item-renderer.st.css';
 
 export const StatefulItemRenderer: React.FC<ListItemProps<ItemData>> = (props) => {
